BOBA/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.27.2
nltk==3.9.1
librosa==0.10.2.post1
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.4
scipy==1.13.1
faster-whisper==1.0.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
